pyproject.toml
[project]
name = "pypi-data"
version = "0.1.0"
description = ""
license = "MIT"
readme = "README.md"
requires-python = "~=3.12.0"
dependencies = [
    "duckdb>=1.0.0",
    "httpx>=0.27.2",
    "more-itertools>=10.4.0",
    "prql-python==0.9.5",
    "psutil>=6.0.0",
    "pyarrow>=17.0.0",
    "pydantic-core>=2.23.4",
    "pydantic>=2.9.2",
    "pygithub",
    "requests",
    "structlog>=24.4.0",
    "tenacity>=9.0.0",
    "tqdm",
    "typer-slim>=0.12.5",
    "zstd>=1.5.5.1",
]

[project.scripts]
pypi-data = "pypi_data.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
